+++ ui/src/lib/auth/types.ts
@@ -7,6 +7,5 @@ export type AuthHook = {
   getAccessToken: () => Promise<string>;
   login: () => void;
   logout: () => void;
-  // TODO: Get the user from the API
   user: User | undefined;
 };
