+++ ui/src/lib/gql/index.ts
@@ -1,6 +1,2 @@
-// TODO: is it okay to export stuff from __gen__
-export * from "./__gen__/fragment-masking";
-export * from "./__gen__/gql";
-export * from "./__gen__/plugins/graphql-request";
 export * from "./queries";
 export * from "./provider/";
