+++ ui/src/lib/gql/queries/user.ts
@@ -1,6 +1,11 @@
-import { graphql } from "@flow/lib/gql";
+import { useQuery } from "@tanstack/react-query";
 
-export const GET_ME = graphql(`
+import { useGraphQLContext } from "@flow/lib/gql";
+
+import { graphql } from "../__gen__";
+
+// Need the queries to build the plugin
+graphql(`
   query GetMe {
     me {
       id
@@ -10,3 +15,14 @@ export const GET_ME = graphql(`
     }
   }
 `);
+
+export const useMeQuery = () => {
+  const graphQLContext = useGraphQLContext();
+  const { data, ...rest } = useQuery({
+    // TODO: Use static keys rather than strings. Export the keys as well
+    queryKey: ["getMe"],
+    queryFn: async () => graphQLContext?.GetMe(),
+  });
+
+  return { data, ...rest };
+};
